seperated theme to seperate folder
